rename Hash component to Truncate component
