+++ src/components/Truncate.vue
